Highly experimental
Trying to change the wheel collider mass to match that of the vessel.
This involves counting the wheel colliders in the vessel and dividing
accordingly. Also needs updating is the parts count changes, as this
indicates the mass has changed. Working on an easy method to check for
updates in vessel mass alone