some small adjustments.
I also fixed the nullref in DustFX. but I also sid a few other things as
well as an attempt at using the water collider detector for something
useful.  WIP of course.  Did a little cleanup here and there too, and
renamed some variables in the DustFX modules to fit better in the usual
variable naming scheme.