DustFX OnDestroy() nullref fix
OnStart() checks is loaded scene is flight before adding the audio
module (as it should) OnDestroy() did no such thing, so when the part
was destroyed in the editor, no instance of the audio effect existed.
I've simplified and added the appropriate checks. Think that's it. I'll
start updating it to be collider hit driven later...