Logging utility is ready, and biome color stuff.
I've merged in the changes done in the DustFX development repository,
and made some adjustments to the camera.  Works flawlessly.  Also
polished up the logging utility I was working on.  I'm not sure it's
actually any better than the standard way, since it requires an object
definition in the file (I tried to make it work without it, but it
refused) and a string passed to the log call that is equal to the
current class (tried to automate that too, but couldn't figure out how
to get the class of the calling method from the log utility, and I
couldn't get the class to be detected automatically in the string in the
file using the log.... C# is annoying sometimes) but, on the bright
side, now the logs will be able to identify themselves with the mod more
clearly.  Note, if no class string name is provided, it will simply not
identify the class that the log is calling from.  It's all pretty
error-proof, and has been tested.  I implemented it into a few of the
other modules as well, as an example of usage.