Trying to get a GUI button to work.  Failing.
The button appears, but won't function properly at all.  I have it
commented out, hoping someone else can take a crack at it.  It's
identical to a button from BDArmory, and it functions fine in that mod,
but in this one is Nullrefs whenever it's clicked, and won't even
execute the commands it's told to.