Just a little tweak...
I noticed that the colors from the camera are a bit more washed out than
I was expecting when testing them out the other day, so I'm just
adjusting their alpha level slightly.  I'm considering that maybe the
alpha level of the color samplers themselves don't need to have any
transparency, and that the dust emitters could handle that themselves,
but that's a future thing to look at.