Merging stuff together.
Okay, while I have yet to test it (too tired of this stuff to try it
now, maybe in a few hours) I believe this is a stable merger between
your changes and my deprecations of a few things we didn't need anymore
and whatnot.  I'm deferring to your new size adjustments and color
corrections, but I still plan to add some variance to the final color of
the dust when I get a chance to look closer at what's been done.  I've
also adjusted the size of the GUI to account for the slider, though it
still needs a label to mark for the end user what it actually does, and
I'd like a readout of what the current value is set to next to or below
the slider itself, or perhaps in the slider label.  We'll work on that
later.