Merging my latest musings and tweakings, plus a few necessary fixes.
Numerous fixes and the addition of a new module I call KFPrecisionMode.

KFPrecisionMode adds a new toggle action group which multiplies the
turning speed & the resource consumption of the wheels on a craft by a
set number which defaults at 2 but can be set in the part config.  In
this way, one can get higher smoothSpeed rates when they want to
precisely align their vessels, but with a cost to go with it to account
for the extra energy required to move the parts that much quicker.
Exact multiplier values are still in need to testing and tweaking.