Latest tweaks and modifications.
Added more xml-documentation to the classes and added a bit to the "#if
debug" checks and/or the log calls to specify the method or the feature
being logged about.  A few renamed variables for clarity, a file name
change ("Utils" to "Extensions"), and a split of the "VesselTools" into
separate class files ("ModuleWaterSlider" and "ModuleCameraShot") for
the two classes that were contained in it for easier development
involving those two classes.  Hope this doesn't mess you up a ton.