Refactoring and more...
Lots of refactoring and some renaming of variables to make more sense to
me.  I've also deprecated the use of the external TextureAnimation
assembly in favor of a rebuild that I did within the KF_plugin source
itself.  I also have a few WIP modules included in this update.  I'm
calling this update a possible 1.9h release, if I do make a release at
this stage.