After a long break, I'm updating this with all my experiments and fixes.
There's a lot of stuff in this update, but very little in functionality
is changed.

Notable changes: TextureAnimator is now integrated into KF_plugin; the
medium wheel has been given a dedicated model that lacks the
repulsorwheel repulsion panel, thus the object destroyer for that part
of the model is totally unnecessary; DustFX has had a few updates but I
cannot, as of yet, reproduce the pause issues and thus cannot seem to
fix them.

No work has been done to make this compatible with the new underwater
stuff in KSP, and the water slider will likely screw that up a ton.
Experiments in implementing a KF specific texture swapper have been
done, but not implemented.

Also began work on a reverse repulsor (attractor?) with a derivative on
"sticky feet" from the critter crawler.  No testing, nor implementation
as of yet.  Lots of reorganization too.